+++ voucher-manager.php
@@ -3,10 +3,10 @@
  * Plugin Name:       Voucher Manager
  * Plugin URI:        https://github.com/mjs512/voucher-manager
  * Description:       Professional One-Time Code Management for WordPress.
- * Version:           1.0.0-rc.1
+ * Version:           1.0.0-rc.2
  * Requires at least: 6.5
  * Requires PHP:      8.1
- * Author:            MJS-Productions e.U.
+ * Author:            MJS-Productions
  * License:           GPL-2.0-or-later
  * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
  * Text Domain:       voucher-manager
@@ -21,7 +21,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'VOUCHER_MANAGER_VERSION', '1.0.0-rc.1' );
+define( 'VOUCHER_MANAGER_VERSION', '1.0.0-rc.2' );
 define( 'VOUCHER_MANAGER_DATABASE_VERSION', '2' );
 define( 'VOUCHER_MANAGER_FILE', __FILE__ );
 define( 'VOUCHER_MANAGER_PATH', plugin_dir_path( __FILE__ ) );